+++ src/editor/DebugMode.h
@@ -16,15 +16,17 @@
 #include <backends/imgui_impl_sdlrenderer2.h>
 #include <SDL2/SDL.h>
 
-class Editor {
-public:
-    Editor();
+#include "../core/SceneManager.h"
 
-    ~Editor();
+class DebugMode {
+public:
+    DebugMode();
+
+    ~DebugMode();
 
     void setup(SDL_Window *window, SDL_Renderer *renderer);
 
-    void update(float deltaTime);
+    void update(float deltaTime, SceneManager &sceneManager);
 
     void render(SDL_Renderer *renderer);
 
@@ -32,17 +34,16 @@ public:
 
     void handleInput(SDL_Event &event);
 
-    void toggleEditor();
-
-    bool isEditorActive() const;
+    void renderEntitiesPanel(SceneManager &sceneManager);
 
 private:
     bool _showDemoWindow = true;
+    bool _showSimpleWindow = true;
     bool _showAnotherWindow = false;
     ImVec4 _clearColor = ImVec4(0.45f, 0.55f, 0.60f, 1.00f);
-    bool _editorActive = false;
 };
 
 #endif //CBIT2D_EDITOR_H
+
 
 #endif // ENABLE_EDITOR