+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Make editor only show when in debug mode
 - Add flag for editor to show/hide and fps to show/hide
+- Add entities component panel in the editor
 
 ### 2024-08-13